Feral Ravager returns
I had originally wanted my ZombieRavager, who is UMA-driven, to replace
the Feral Ravager. However it appears that I cannot change the movement
speed of a crawling UMA zombie (this may also apply to walking ones).

As such I found him to be way too slow to be a threat and it removed the
entire point of his character. Thus I have brought back the feral.

However to keep the ZombieFeral I created I have now removed the feral
spider zombie and and have set the ZombieFeral to fill his role instead.